+++ source/Core/HELM/Vector.h
@@ -4,12 +4,12 @@ template<class T>
 class Vector
 {
 public:
-	Vector(size_t n);
+	Vector(int n);
 	Vector(Vector<T> const &rhs);
 	~Vector();
 
-	size_t getCount() const;
-	void set(size_t i, T const &value);
+	int getCount() const;
+	void set(int i, T const &value);
 	T dot(Vector<T> const &rhs) const;
 	T squaredNorm() const;
 	void weightedSum(Vector<T> const &x, T const &yWeight, Vector<T> const &y);
@@ -18,7 +18,7 @@ public:
 	void subtract(Vector<T> const &x, Vector<T> const &y);
 	void conjugate();
 
-	T const& operator()(size_t i) const;
+	T const& operator()(int i) const;
 	Vector<T> const& operator=(Vector<T> const &rhs);
 
 private:
@@ -28,7 +28,7 @@ private:
 	void setToZero();
 
 private:
-	const size_t _count;
+	const int _count;
 	T *_values;
 };
